# Conflicts:
#	.flutter-plugins-dependencies
#	assets/flutter_i18n/en.json
#	assets/flutter_i18n/ru.json
#	assets/flutter_i18n/tr.json
#	build_variant.yaml
#	lib/background/background_sync.dart
#	lib/main.dart
#	lib/modules/auth/blocs/auth_bloc/auth_bloc.dart
#	lib/modules/auth/blocs/auth_bloc/auth_methods.dart
#	pubspec.lock
#	pubspec.yaml
#	pubspec.yaml.yaml

+++ lib/modules/auth/blocs/auth_bloc/auth_methods.dart
@@ -118,13 +118,15 @@ class AuthMethods {
     return _accountsDao.getAccounts(user.localId);
   }
 
-  Future<void> logout(User user) async {
+  Future<void> logout(int currentUserId, User user) async {
     _authApi.logout(user);
 
     final futures = [
       deleteUserRelatedData(user),
+      if(user.localId==currentUserId)
       _cryptoStorage.deleteAll(),
       _authLocal.deleteSelectedUserLocalId(),
+      if(user.localId==currentUserId)
       _authLocal.deleteSelectedAccountId(),
       _usersDao.deleteUser(user.localId),
       _accountsDao.deleteAccountsOfUser(user.localId),
